% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeneNames.R
\name{addGeneNames}
\alias{addGeneNames}
\title{Add Gene Names}
\usage{
addGeneNames(x, t2g = NULL)
}
\arguments{
\item{x}{Expression matrix (or any matrix) that has ENSG IDs as row.names}

\item{t2g}{Transcript to Genename table for mapping (if available). If none is supplied then will generate one from BioMart (takes longer).}
}
\value{
Data table with columns for gene names and ENSG IDs.
}
\description{
Just a quick way to convert gene matrices with ENSGs as row.names to data tables with columns for human readable gene names and ENSGs.
}
