% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allDuplicated.R
\name{allDuplicated}
\alias{allDuplicated}
\title{All Duplicated
Return logical vector marking any value that has a duplication of another value.
 Not just the duplicates of smaller indexes.}
\usage{
allDuplicated(vec)
}
\arguments{
\item{vec}{Input vector that might have duplicate values}
}
\value{

}
\description{
All Duplicated
Return logical vector marking any value that has a duplication of another value.
 Not just the duplicates of smaller indexes.
}
