% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.rats.R
\name{auto.rats}
\alias{auto.rats}
\title{Autosomal Ratios (auto.rats)}
\usage{
auto.rats(typ.obj, color.by = NULL)
}
\arguments{
\item{typ.obj}{typ objects of type 'typ.gene' or 'typ.trans'}

\item{color.by}{OPTIONAL Specify column in object meta data
to color bar plots (i.e. "sex" or "karyotype")}
}
\value{
matrix of ratios, pdf containing plots
}
\description{
For each sample, calculates ratio of normalized
   expression for all measurements from each autosome to all
   other autosomes (i.e. autosome ratio). Then compares ratios
   across all samples for each chromosome. Can detect severe
   chromosomal abnormalities (i.e. chromosomal duplication or loss).
   Outputs dataframe of all ratios with column for each chromosome
   and a pdf containing all ratio plots for chr1:chr22. Pdfs are
   writen to a subfolder named 'plots'
}
