% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenRichment.R
\name{BenRichment}
\alias{BenRichment}
\title{Gene Ontology Enrichment on data frame of gene symbols}
\usage{
BenRichment(x)
}
\arguments{
\item{x}{A dataframe of gene lists in columns. Column names = name of gene list.
Genes should be standardized gene symbols}
}
\description{
Gene Ontology Enrichment on data frame of gene symbols
}
