% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chop.R
\name{chop}
\alias{chop}
\title{Filter gene matrices based on k over A}
\usage{
chop(x, k = NULL, A = NULL)
}
\arguments{
\item{x}{matrix to be filtered with measurements in rows and samples in columns}

\item{k}{number of samples that need to be higher than A to pass filter}

\item{A}{minimum value to pass filter}
}
\value{
filtered matrix
}
\description{
This is a slightly easier and shorter function to run the k over A algorithm.
Filters matrices so that output only includes measurements (genes, proteins, etc.)
higher than A in at least k samples.
}
