% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanTable.R
\name{cleanTable}
\alias{cleanTable}
\title{Clean limma test table}
\usage{
cleanTable(x, lfc.cut = 0.5, adj.p.cut = 0.05, map = map)
}
\arguments{
\item{x}{test table output from limma topTable function}

\item{lfc.cut}{minimum  absolute log fold change for significant hit call. Default = 0.5}

\item{adj.p.cut}{adjusted P value cutoff for significant hit call. Default = .05}

\item{map}{map that connects ENSG values (or other row.names) to common gene names}
}
\description{
Takes a test table outputted from limma topTable function and generates table that
can be easily used for ggplot functions. Can be directly input to the Vulcan function
for plotting volcano plots. Assumes that row.names of x are ENSG IDs. Those ENSG IDs are
mapped using the "map" file with column names "ensembl_gene_id" and "gene" for ENSG ID and
gene name respectively.
}
