% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTT.R
\name{cleanTT}
\alias{cleanTT}
\title{Clean test table
cleans up the sleuth_results() test table output}
\usage{
cleanTT(x)
}
\arguments{
\item{x}{test table from sleuth_results() function}
}
\value{

}
\description{
Clean test table
cleans up the sleuth_results() test table output
}
