% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompPairwiseCorMat.R
\name{CompPairwiseCorMat}
\alias{CompPairwiseCorMat}
\title{Comprehensive test of Pair-wise Correlation Parameters}
\usage{
CompPairwiseCorMat(x, y)
}
\arguments{
\item{x}{matrix 1}

\item{y}{matrix 2 (see pairwise CorMat for guidence)}
}
\value{
Data frame of skews for different adjustments
}
\description{
Run pairwiseCorMat.R function varying all parameters to
test the effect on skew
}
