% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGsByCluster.R
\name{DEbyClust}
\alias{DEbyClust}
\title{Cluster-wise test for DEGs of Seurat Object}
\usage{
DEbyClust(x, group = "seurat_clusters", comp = NULL)
}
\arguments{
\item{x}{Seurat object with clusters defined}

\item{group}{Category to compare with ("seurat_clusters" is default). Can be
any parameter specified in meta.data column that segregated cells into labeled groups.}

\item{comp}{Comparison to be tested. Must be specified and be a column in meta.data.
Use quoted column header of comparison, i.e. comp = "treatment" where "treatment" is a
column header for cells that were "stim." or "vehicle".}
}
\description{
For analysis within each cluster (or other set of cell group defined in meta.data)
for DEGs (markers) that are significant between specified comparison.
Comparison must exist as a column in seuratObject(at)meta.data" column with
two possible values (i.e. ko, wt, or treated, sham)
}
