% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{filt}
\alias{filt}
\title{Filter Data}
\usage{
filt(x, k = NULL, A = NULL)
}
\arguments{
\item{x}{Input expression matrix}

\item{k}{Number of samples required to have values over that specified in A}

\item{A}{Target value above which is required in k samples to be retained}
}
\value{
K over A filtered data
}
\description{
Just a wrapper for k over A filtering with k being the number of samples
with values over A required to not be filtered out.
}
