% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCN.R
\name{gcn}
\alias{gcn}
\title{Gene Concept Network Plot
Simple generator of gene concept network plots using enrichment object}
\usage{
gcn(x, nodes = 5, label.size = "medium", circle = FALSE, color.lines = TRUE)
}
\arguments{
\item{x}{enrichment object output from running mkEnrich}

\item{nodes}{number of nodes (enriched pathways) to plot. default is 4}

\item{label.size}{How large to make the gene labels (none, teeny, small, medium, large)}

\item{circle}{Logical. Should the plot be constrained to a circle (default=FALSE)}

\item{color.lines}{Logical. Should the lines be colored by gene set? (default=FALSE)}
}
\value{

}
\description{
Gene Concept Network Plot
Simple generator of gene concept network plots using enrichment object
}
