% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnsGeneMaps.R
\name{getEnsGeneMaps}
\alias{getEnsGeneMaps}
\title{Get Current Ensembl Gene Maps}
\usage{
getEnsGeneMaps()
}
\value{

}
\description{
Requires no input. Returns a list object with 2 data frames for mapping ENSGs and ENSTs to gene symbols.
 For use with human genomic data only.

 tgn data frame has columns for
 transcript ID (ENST numbers) labeled "target_id"
 gene ID (ENSG numbers) labeled "ens_gene"
 gene symbols (human readable gene symbols) labeled "ext_gene

 gn contains only ens_gene, and ext_gene columns (unique)
}
