% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2mCorrPlot.R
\name{m2mCorrPlot}
\alias{m2mCorrPlot}
\title{Dot matrix of Correlations}
\usage{
m2mCorrPlot(mat1, mat2, method = "pearson")
}
\arguments{
\item{mat1}{matrix one with samples in rows and measurements in columns}

\item{mat2}{optional: matrix of measures using overlapping samples as matrix 1. Samples not appearing in both
matrices will be culled. mat2 will be transposed if necessary.}

\item{method}{Type of correlation to perform "pearson" or "spearman"}
}
\description{
Generates a matrix plot of graphical representations of correlations using either 1 or 2 matrices.
This is a wrapper for the corrplot function (from the CorrPlot package) that adds the step of combining
matrices into one (if necessary) and calculating correlations for plotting. The problem of graphing larger
matrices still remains. This is generally good up to 10 columns of measures after which the graphics can get
ugly. Generally good to minimize the length of measurement (column) names.
}
