% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.typ.R
\name{make.typ}
\alias{make.typ}
\title{Make Standardized 'typ' object}
\usage{
make.typ(x, type = "gene", meta = NULL, regress = NULL)
}
\arguments{
\item{x}{sleuth object or expression matrix}

\item{type}{c("gene", "trans", "prot", "deg", "dep")}

\item{meta}{meta data information as data frame (not required if sleuth obj passed)}

\item{regress}{OPTIONAL column names in meta data file specifying regression}
}
\value{
object of class 'typ'
}
\description{
Generates uniform R object of class 'typ' for use in
   downstream processes
}
