% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRank.R
\name{makeRank}
\alias{makeRank}
\title{Make Rank File for GSEA analysis}
\usage{
makeRank(x, fileName, sig, direc)
}
\arguments{
\item{x}{Test table with ENSGs as row.names and columns for significance
of test and direction of change
Designed to directly convert limma topTable.}

\item{fileName}{specify name of file to be written including .rnk suffix.
(i.e. file = "output.rnk")}

\item{sig}{name of column containing significance, not quoted}

\item{direc}{name of column containing directionality. b value for DEG or r value for correlation}
}
\value{

}
\description{
Takes comparison table with ENSG IDs as row.names and columns for
significance and for direction
Ranks genes using -log10(sig) * dir
}
