% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMprep.R
\name{NAMprep}
\alias{NAMprep}
\title{NAM prep
Running the NAMscore function on a Seurat object uses the scale.data slot
If Scale() was only run on the most variable features (default) rather than
all features then many of the key markers could be lost (especially if one
cell type is over-represented). NAM prep process a Seurat object to run the
ScaleData() function on all features rather than only the 2000 most variable.
Can take a while.}
\usage{
NAMprep(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{

}
\description{
NAM prep
Running the NAMscore function on a Seurat object uses the scale.data slot
If Scale() was only run on the most variable features (default) rather than
all features then many of the key markers could be lost (especially if one
cell type is over-represented). NAM prep process a Seurat object to run the
ScaleData() function on all features rather than only the 2000 most variable.
Can take a while.
}
