% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noNANA.R
\name{noNANA}
\alias{noNANA}
\title{Removes rows with any NA values}
\usage{
noNANA(x)
}
\arguments{
\item{x}{matrix with NA values}
}
\value{

}
\description{
This is just a quicker function to remove any row containing any
NA value. Note this may be overkill in many cases. I use it for
processing proteomic matrices where NA is an indicator of unreliable
measurements.
}
