% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noNANAtt.R
\name{noNANAtt}
\alias{noNANAtt}
\title{Sleuth Test table filter, removes NA rows and unmapped genes}
\usage{
noNANAtt(x)
}
\arguments{
\item{x}{sleuth test table from wald test}
}
\value{

}
\description{
Sleuth Test table filter, removes NA rows and unmapped genes
}
