% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NULLdistPWC.R
\name{NULLdistPWC}
\alias{NULLdistPWC}
\title{The Null Distribution generator for pairwiseCorMat}
\usage{
NULLdistPWC(x, y, method = "spearman", iter = 1000, pct = 1)
}
\arguments{
\item{x}{matrix 1}

\item{y}{matrix 2}

\item{method}{c("spearman", "pearson")}

\item{iter}{number of iterations (over 1000 and this gets very slow)}

\item{pct}{percentage of upper variance genes to test}
}
\description{
Generates a Null distribution and tests significance of skew seen in pairwiseCorMat
}
