% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds2of3.R
\name{odds2of3}
\alias{odds2of3}
\title{Odds of passing 2 out of 3 Comparisons}
\usage{
odds2of3(x, y, method = "pearson", iter = 100, p.cutoff = 0.05, pct = 1)
}
\arguments{
\item{x}{matrix of 3 columns to compare (could make this extensible but haven't yet)}

\item{y}{matrix of expression values (or proteomics or anything really)}

\item{method}{correlation method to use "spearman", "pearson"}

\item{iter}{number of iterations (can get slow if too large)}

\item{p.cutoff}{p value cutoff for determining significance}

\item{pct}{pct of upper variance to include in comparisons (default = 1 or include all)}
}
\description{
Highly specialized function to compare whether a set of values in matrix y, are significantly correlated (uncorrected pVal)
to 2 out of the 3 columns of values supplied in matrix x. Runs through each column in y and correlates with each
of the columns in x then runs a set of tests to determine if correlation passes set rval (set to .35), that
rvalue signs agree, and finally that 2 out of the 3 pvals exceed the cutoff (p.cutoff). Subsequently runs a NULL distribution
analysis to determine how frequently the observed number of passers would be expected to occur by chance (FDR)
}
