% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseCorMat.R
\name{pairwiseCorMat}
\alias{pairwiseCorMat}
\title{Pairwise correlation matrix}
\usage{
pairwiseCorMat(x, y, method = "pearson", pct = "1.0")
}
\arguments{
\item{x}{matrix 1}

\item{y}{matrix 2}

\item{method}{"pearson" or "spearman"}

\item{pct}{percentage of upper variance to include}
}
\description{
Temporary description
}
