% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseCorSkewNULL.R
\name{pairwiseCorSkewNULL}
\alias{pairwiseCorSkewNULL}
\title{Pairwise Correlation Skew NULL Distribution (pairwiseCorSkewNULL)}
\usage{
pairwiseCorSkewNULL(x, y, method = "spearman", iter = 1000, pct = 1)
}
\arguments{
\item{x}{matrix 1}

\item{y}{matrix 2}

\item{method}{Type of correlation to perform "spearman" or "pearson"}

\item{iter}{Number of iterations}

\item{pct}{percentage of upper variance to include}
}
\description{
Generates a NULL distribution for "pairwiseCorSkew" to test how frequently
one expects to see an observed skew by chance. Note that this can take a
while to run if inputing large matrices with high numbers of iterations.
It's a good idea to test with 100 iterations first before ramping up.
After generation x <- pairwiseCorSkewNULL, the distribution can be visualized
by plotting plot(density(x)), then abline(v = observedSkew from pairwiseCorSkew)
}
