% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platelabs.R
\name{platelabs}
\alias{platelabs}
\title{Plate labels}
\usage{
platelabs(plate = 96)
}
\arguments{
\item{plate}{Enter the plate type as number of wells (i.e. 6, 12, 24, 48, 96, 384)}
}
\description{
Builds a set of labels for diffent tissue culture plate sizes (i.e. A1, A2...). Not rocket science.
}
