% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_composition.R
\name{plot_cluster_composition}
\alias{plot_cluster_composition}
\title{Plot Cluster Composition}
\usage{
plot_cluster_composition(
  so,
  cluster = NULL,
  comp.report = NULL,
  table.out = FALSE
)
}
\arguments{
\item{so}{a Seurat object}

\item{cluster}{Cluster number or ID to analyse}

\item{comp.report}{name of the metadata parameter to report composition}

\item{table.out}{LOGICAL. Should only table values be reported}
}
\description{
Takes a Seurat object as input and uses embedded metadata to generate a
set of stacked bar plots illustrating compositional percentages
}
