% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantNorm.R
\name{quantNorm}
\alias{quantNorm}
\title{Quantile Normalization of Expression (RNAseq) data}
\usage{
quantNorm(x)
}
\arguments{
\item{x}{Matrix to be quantile normalized}
}
\description{
Just a simple wrapper for preprocessCore() which fixes the problem of removed colnames and row.names
during execution
}
