% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickPCA.R
\name{quickPCA}
\alias{quickPCA}
\title{Quick PCA}
\usage{
quickPCA(
  x,
  meta = NULL,
  sample = "sample",
  scale = FALSE,
  colorBy = NULL,
  samp.labels = FALSE,
  plot.title = NULL
)
}
\arguments{
\item{x}{Your expression matrix with samples in columns and measurements (genes/proteins/etc.) in rows.}

\item{meta}{Metadata object, ideally with column labeled "sample" that matches the names of the columns in your expression matrix. all(colnames(x) == meta[,sample]) must return TRUE}

\item{sample}{Name of metadata column with sample names. Default is "sample" specify if yours is different (i.e. "Sample", "samp." or some other foolish name)}

\item{scale}{(Logical) Whether or not prcomp should scale the data prior to running PCA. Leave FALSE if data is already scaled/log transformed. Set to TRUE for raw data values}

\item{colorBy}{(Optional) Name of the column to color dots by. (i.e. "treatment", "batch", "sex")}

\item{samp.labels}{(Logical) Should individual points be}

\item{plot.title}{(Optional) Title for your PCA plot.}
}
\value{
PCA plot
}
\description{
Just an abbreviated wrapper for prcomp() and fviz_pca_ind because I get tired of all the lines of code. Cleans up scripts with many PCA plots.
}
