% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCompare.R
\name{rCompare}
\alias{rCompare}
\title{Overlay Waterfall plot of pos and negative correlations
Takes output of pairwiseCorrSkew() and generates an inverted
waterfall plot visually comparing pos correlations, FDR passers
and neg correlations. Plot is outputted as a pdf}
\usage{
rCompare(x, title = "add title", pdfName = "output.pdf")
}
\arguments{
\item{x}{Output list object from running pairwiseCorrSkew}

\item{title}{Title of plot (describe the comparison). I.e. "iAstro to Brain"}

\item{pdfName}{Quoted name of the pdf file to generate as output}
}
\description{
Overlay Waterfall plot of pos and negative correlations
Takes output of pairwiseCorrSkew() and generates an inverted
waterfall plot visually comparing pos correlations, FDR passers
and neg correlations. Plot is outputted as a pdf
}
