% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcorradj.R
\name{rcorradj}
\alias{rcorradj}
\title{rcorr with adjusted P values included}
\usage{
rcorradj(x, y, type = c("pearson", "spearman"), p.adj = "BH")
}
\arguments{
\item{x}{a numeric matrix with at least 5 rows and at least 2
columns (if y is absent). For print, x is an object produced
by rcorr.}

\item{y}{a numeric vector or matrix which will be concatenated to x.
If y is omitted for rcorr, x must be a matrix.}

\item{type}{specifies the type of correlations to compute. Spearman correlations are the Pearson linear correlations computed on the ranks of non-missing elements, using midranks for ties.}

\item{p.adj}{Method to be used for p value adjustment "BH" default}
}
\value{

}
\description{
This is just Hmisc's rcorr function with two changes
1) NA in diagonal are replaced with 0's
2) returns adjusted pvalues in list object
}
