% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportProgress.R
\name{reportProgress}
\alias{reportProgress}
\title{Report Progress}
\usage{
reportProgress(n, iter, t1)
}
\arguments{
\item{n}{current iteration}

\item{iter}{total number of iterations to be run}

\item{t1}{starting time}
}
\value{

}
\description{
For use within long running scripts. Will report progress in 10% intervals when running iterations.
Must start script by defining t1 <- Sys.time()
}
