% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.sleuth.R
\name{run.sleuth}
\alias{run.sleuth}
\title{Run sleuth on kallisto output}
\usage{
run.sleuth(in.dir = "results", meta = "meta.csv", level = c("gene"))
}
\arguments{
\item{in.dir}{Name of directory with kallisto output. Sub direstories should be structured as:
sampleName1/{abundance.h5,abundance.tsv,run_info.json}
sampleName2/{abundance.h5,abundance.tsv,run_info.json}
...}

\item{meta}{Name of meta data file in working directory i.e. "AH1-18meta.csv"}

\item{level}{Level of analysis to perform "gene" or "trans". Default is both.}
}
\value{
sleuth object(s)
}
\description{
For somewhat easier processing of kallisto output. If you have your kallisto output files
in a sub-directory named 'results' and a matching metadata file named 'meta' then you only
need to execute 'output <- run.sleuth() to output both gene-level and transcript-level
sleuth results (returned as a list). Can specify  custom directory or metadata file names,
and specify "gene" or "transcript" levels in arguments.
}
