% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seahorseImport.R
\name{seahorseImport}
\alias{seahorseImport}
\title{Seahorse Data import}
\usage{
seahorseImport(file = NULL)
}
\arguments{
\item{file}{quoted name of .csv file to import from above process}
}
\value{

}
\description{
From mito stress test report (.xlsm) which is generated from the Wave software,
 copy and transpose average assay parameters calculations from the "Measures sheet"
 (generally starts in cell B4). Copy all measurement data up to ATP Production.
 Transpose data to new workbook so that sample names are in rows (merged cells) and
 measurements are in columns. Save as .csv file. This is the input for the seahorseImport()
}
