% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sniff.R
\name{sniff}
\alias{sniff}
\title{Quick peek (sniff) of typ objects}
\usage{
sniff(typ.obj, what = "meta")
}
\arguments{
\item{typ.obj}{object created with typGumbo::make.typ() function}

\item{what}{name of field to display. Default is what = "meta" options are "gene" (snippet
of gene-level matris), "trans" (snippet of transcription-level matrix), "info"
(processing information), "all" (all of the above).}
}
\value{
Quick View
}
\description{
Just a quick peek. Nothing complicated
}
