% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stickem.R
\name{stickem}
\alias{stickem}
\title{Combine two numeric matrices with overlapping samples}
\usage{
stickem(mat1, mat2)
}
\arguments{
\item{mat1}{matrix 1 containing sample names as either row.names or colnames.
Orientation of mat1 will be the orientation of the output matrix.}

\item{mat2}{matrix 2 containing samples names as either row.names or colnames}
}
\value{

}
\description{
Two matrices with overlapping samples and different measurements are combined. Matrix supplied as mat1
is used to orient final output. Sample names must have identical formats between the two matrices and must have
uniquely identifiable names (i.e. "BR04" rather than vague sample names like "sample1"). That can get you into trouble
Samples can be in either row.names or colnames on either matrix. The script will transpose mat2 if necessary.
Samples not in both matrices will be removed and output will be in the same orientation (samples in rows or columns) as mat1.
}
