% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.smooth.R
\name{syn.smooth}
\alias{syn.smooth}
\title{Synaptic Smoothing pipeline}
\usage{
syn.smooth(in.dir, csv = TRUE, xlsx = TRUE, filter = TRUE)
}
\arguments{
\item{in.dir}{directory where all input .csv files are located. Omit trailing slash.
i.e. in.dir = "LViN15", or in.dir = "LViN15/results"}

\item{csv}{Logical. Indicate whether you want to write .csv files to output directory}

\item{xlsx}{Logical. Indicate whether you want to write .xlsx file with separate sheets
for each output matrix to output directory. NOTE: this uses the xlsx::write.xlsx function
which takes a lot of time and memory and should not be used for very large data sets (larger than
1000 observations)}

\item{filter}{logical. indicating whether data should be filtered to exclude measurements with pctFC
less than 2. Even with this filter off the data will be filtered to remove measurements with pctDeltaF
greater than 0 (i.e. requires positive fluorescence change)}
}
\description{
A highly specialized pipeline for processing synaptic vesicle release assay results.
It takes in a directory containing .csv files in a very specific TPP format (Tracy's Proprietary Processing)
and returns a set of processed csvs for each assay along with an excel file containing a
summary of the results and some analytics. Input csv's from excel file should have 5 header metrics rows from row 1 they
must be 1) F0 (average starting fluor), 2) Fstim (average stim fluor), 3) Fun (max unquenched fluro), 4) FC_Fun-F0 (fold change of
unquenched max relative to baseline(F0)), and 5) deltaF (increase in fluor in stimulated vs. F0). First column (under metrics labels)
must be a space (extra column) for time values. The script will remove this column and if it contains data rather than
timestamps then the data will be lost. The worksheets used to generate the csv's will often have extra columns at the end of the sheet.
They start at row 6 (after the header metrics) and report average and error fluor across all measurements for a given time. There can be
2-4 of these or none. It doesn't matter as any column after the last value in the row 1 metrics will be removed.
}
\details{
If this function returns a Java out of memory heap space error then try restarting R and then run the following line
  before loading typGumbo

  options(java.parameters = c("-XX:+UseConcMarkSweepGC", "-Xmx8192m"))
}
