% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_smooth_1color.R
\name{syn_smooth_1color}
\alias{syn_smooth_1color}
\title{Synaptic Vesical Release Assay Smoothing (1 Cell type)}
\usage{
syn_smooth_1color(in_dir, csv = TRUE, xlsx = TRUE, filter = TRUE)
}
\arguments{
\item{in_dir}{Quoted path to directory with *.xlsx file}

\item{csv}{LOGICAL Do you want to output *.csv files. Default is TRUE.}

\item{xlsx}{LOGICAL Do you want to output *.xlsx files. Default is TRUE.}

\item{filter}{LOGICAL Do you want to apply signal filter}
}
\value{
Writes summary files for downstream usage in Prism or excel.
Files are organized in hierarchical directory related to sampleIDs.
}
\description{
Synaptic Vesical Release Assay Smoothing (1 Cell type)
}
