% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_smooth_2color.R
\name{syn_smooth_2color}
\alias{syn_smooth_2color}
\title{Synaptic Vesical Release Assay Smoothing (2 Cell types)}
\usage{
syn_smooth_2color(
  in_dir,
  csv = TRUE,
  xlsx = TRUE,
  filter = TRUE,
  split_by = "RFP",
  upper_pct_cut = 33
)
}
\arguments{
\item{in_dir}{Quoted path to directory with *.xlsx file}

\item{csv}{LOGICAL Do you want to output *.csv files. Default is TRUE.}

\item{xlsx}{LOGICAL Do you want to output *.xlsx files. Default is TRUE.}

\item{filter}{LOGICAL Do you want to apply signal filter}

\item{split_by}{CHARACTER What is the name of the signal in time point 1
that defines the second cell type. Default is "RFP"}

\item{upper_pct_cut}{INTEGER upper percentage of ROIs to call positive
for the split_by signal. Default is 33}
}
\value{
Writes summary files for downstream usage in Prism or excel.
Files are organized in hierarchical directory related to sampleIDs.
}
\description{
Synaptic Vesical Release Assay Smoothing (2 Cell types)
}
