% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrimEnrich.R
\name{TrimEnrich}
\alias{TrimEnrich}
\title{Trim mkEnrich() results}
\usage{
TrimEnrich(x, modules = NULL)
}
\arguments{
\item{x}{Output object from mkEnrich() function}

\item{modules}{Row numbers (indices) of pathways to keep}
}
\description{
Many of the enriched pathways from GO enrichment are highly overlapping.
i.e. BC: Ribosome and BP Ribosome Function, etc.
To clean up the downstream Gene concept networks it can be helpful to trim
redundant enriched pathways.
}
