% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typ.filt.R
\name{typ.filt}
\alias{typ.filt}
\title{Filter typ Object Expression Data}
\usage{
typ.filt(
  typ.obj,
  type = c("k.over.a", "min.mean", "top.x"),
  k = NULL,
  a = NULL,
  min = NULL,
  x = NULL
)
}
\arguments{
\item{typ.obj}{expression object created with 'make.typ()' function}

\item{type}{c("k.over.a","min.mean", ""top.x") can apply multiple}

\item{k}{number of samples required to have 'a' expression level}

\item{a}{expression level required}

\item{min}{minimum average expression}

\item{x}{number of top measurements to return}
}
\value{
typ object same as entered but with filtered data and info
        detailing the filtering parameters
}
\description{
Custom filtering of typ objects using either
k over a, minimum average expression, or highest x measurements
(based on average expression)
}
