% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typ.prot.map.R
\name{typ.prot.map}
\alias{typ.prot.map}
\title{TYP Proteomics Map
Attaches genomic mapping information to proteomics map. Generally used as an internal function in typGumbo.}
\usage{
typ.prot.map(prot.map)
}
\arguments{
\item{prot.map}{A simple mapping table derived from the proteomics processing that contains a uniprot ID labeled 'prot'
and a gene name labeled 'gene'.}
}
\description{
TYP Proteomics Map
Attaches genomic mapping information to proteomics map. Generally used as an internal function in typGumbo.
}
