% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verOut.R
\name{verOut}
\alias{verOut}
\title{Version Out for removing ENSG and ENST version extensions}
\usage{
verOut(x)
}
\arguments{
\item{x}{a vector of transcripts(ENSTs) or genes(ENSGs) with versions}
}
\description{
In some gene or transcript expression data there are trailing numbers
    after a decimal indicating the version of that sequence originally
    matched against. E.g. ENSG000000876.5  - these can cause problems
    with mapping to gene meta data and are often best removed for downstream
    analysis
}
