% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewEnrich.R
\name{ViewEnrich}
\alias{ViewEnrich}
\title{View GO Enrichment Results
  simple viewer of significantly enriched pathways that are output from mkEnrich()}
\usage{
ViewEnrich(x)
}
\arguments{
\item{x}{Output object from mkEnrich()}
}
\value{
Streamlined version of results from mkEnrich object
}
\description{
View GO Enrichment Results
  simple viewer of significantly enriched pathways that are output from mkEnrich()
}
