% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulcan.R
\name{vulcan}
\alias{vulcan}
\title{Vulcan: Generate easy volcano plots from a "clean" test table}
\usage{
vulcan(x, labels = TRUE, repel = 1, pad = 0.25, labelsize = 4, title)
}
\arguments{
\item{x}{Output from CleanTable (from limma) or equivalent}

\item{labels}{LOGICAL. Should significant genes be labeled.}

\item{repel}{Force of point label repulsion (from geom_text_repel()). Default = 1}

\item{pad}{Padding around label boxes (from geom_text_repel()). Default = 0.25}

\item{labelsize}{Size of labels (from geom_text_repel()). Default = 4}

\item{title}{Title for volcano plot. Default is no title.}
}
\value{
volcano plot
}
\description{
Takes a clean test table and generates a volcano plot with relative ease
and includes a few key adjustable parameters. Test table is trimmed of all rows with
NAs (retained in sleuth_results()). ENSGs without mapped ext_gene names are also trimmed.
}
