% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTT.R
\name{cleanTT}
\alias{cleanTT}
\title{Clean test table}
\usage{
cleanTT(x, bcut = 0.5, qcut = 0.05)
}
\arguments{
\item{x}{test table from sleuth_results() function}

\item{bcut}{cutoff value for fold change. Default is .5 or 50% change}

\item{qcut}{cutoff value for significance. Default is q value of .05}
}
\value{

}
\description{
cleans up the sleuth_results() test table output
}
