% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTT.R
\name{cleanTT}
\alias{cleanTT}
\title{Clean Sleuth Test Table}
\usage{
cleanTT(x, bcut = 0.5, qcut = 0.05)
}
\arguments{
\item{x}{Test table from sleuth_results}

\item{bcut}{Cutoff for fold change (z score) significance}

\item{qcut}{Cutoff for qvalue (FDR) significance}
}
\value{
cleaned test table
}
\description{
Cleans up test table output from sleuth_results.
}
