% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanXM.R
\name{cleanXM}
\alias{cleanXM}
\title{Expression matrix cleaning tool}
\usage{
cleanXM(x, k = 0, A = 0)
}
\arguments{
\item{x}{Input expression matrix, measurements in rows}

\item{k}{Number of samples whose expression need to be higher than A}

\item{A}{Expression (measurement) level to target}
}
\value{

}
\description{
This is a short pipeline that filters expression matrices based on k over A
arguments, log2() transforms and quantile normalizes. Unnecessary for ratiometric
data or previously processed data but useful for raw or TPM matrices.
Function will remove all measurements with 0 values
in all samples (zero variance). It will test matrix for asymmetric negative values (often resulting from
regressing batch effects) and raise all values by abs(smallest negative value) to remove any negative
expression measurement.
}
