% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorVMatSig.R
\name{CorVMatSig}
\alias{CorVMatSig}
\title{Correlate vector to matrix and test significance}
\usage{
CorVMatSig(x, y, method = "pearson", iter = 100, p.cutoff = 0.05, pct = 1)
}
\arguments{
\item{x}{Named Vector to be correlated with matrix y}

\item{y}{Matrix with compatible names as vector x}

\item{method}{type of correlation to use "pearson" or "spearman"}

\item{iter}{number of iterations to run}

\item{p.cutoff}{p value cutoff for determining significance}

\item{pct}{percentage upper variance to use}
}
\value{

}
\description{
Correlate Vector to rows in a  matrix and determine significance against a NULL distribution. Input vector (x)
 is a named vector that corresponds to row.names in matrix (y). Correlations between
}
