% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnsGeneMaps.R
\name{getEnsGeneMaps}
\alias{getEnsGeneMaps}
\title{Get Current Ensembl Gene Maps}
\usage{
getEnsGeneMaps(type = c("full", "t2g", "t2n", "g2n", "g2e", "n2e"))
}
\arguments{
\item{type}{Indicate type of map desired. c("full", "t2g", "t2n", "g2n", "g2e", "n2e")
"full" = full map containing Ensemble transcript IDs (ENSTs), gene IDs (ENSGs), external gene names (symbols), and Entrez IDs (####)
  full map will contain NAs where identifiers don't have mapped counterparts
"t2g" = ENST and ENSG
"t2n" = ENST and gene names (symbols)
"g2n" = ENSGs and gene names (symbols)
"g2e" = ENSGs and Entrez IDs
"n2e" = gene names (symbols) and Entrez IDs}
}
\value{

}
\description{
Returns a data frame for mapping specified genetic identifiers.
 For use with human data only.
}
