% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkEnrich.R
\name{mkEnrich}
\alias{mkEnrich}
\title{Make Gene Ontology Enrichment Object
Simple script that uses clusterProfiler to generate GO enrichment object
Outputted object can be fed into downstream functions like gene concept network plots
Can take a bit of time to run}
\usage{
mkEnrich(x, qvalueCutoff = 0.01)
}
\arguments{
\item{x}{quoted path to *.csv file containing ONE column of gene names}

\item{qvalueCutoff}{Highest qvalue returned (default = .01)}
}
\value{
GO enrichment object
}
\description{
Make Gene Ontology Enrichment Object
Simple script that uses clusterProfiler to generate GO enrichment object
Outputted object can be fed into downstream functions like gene concept network plots
Can take a bit of time to run
}
