% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMscore.R
\name{NAMscore}
\alias{NAMscore}
\title{NAM (iN, iAst, iMGL) Score Generator}
\usage{
NAMscore(object)
}
\arguments{
\item{object}{A seurat object}
}
\value{

}
\description{
Used for assigning cells to original culture when they've been mixed
(i.e. in triple cultures or in combined monocultures). REQUIRES that
ScaleData() has been applied to ALL features (not just most variable).
NAMprep() function can be used for that purpose if necessary.
NAMscore assigns an iN, iAst, and a iMGL score value for each cell.
Input is a Seurat object and output is a Seurat object with altered metadata
that includes 4 new columns. The 3 scores (iN.score, iAst.score, and iMGL.score)
and a "NAM.bin" column that assigns each cell to predicted culture (iN, iAst, or iMGL)
or as "unk". Returned Seurat object has "NAM.bin" set as active.ident
So running DimPlot(object) will display active reduction (probably UMAP)
colored by NAM.bin
Reset active.ident to "seurat_clusters" to display cluster info on UMAP
i.e.
object <- SetIdent(object, value = "seurat_clusters")
}
\details{
Note that the markers for iNs, iAst, and iMGLs have been hard coded and are derived from
scRNAseq of independent iAst, iNs, and iMGL cultures where identities were overtly known.
Strong Positives were defined be being >= 90% of cell type of interest and <= 10% of other cells.
}
