% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noNANA.R, R/noNANAtt.R
\name{NoNANA}
\alias{NoNANA}
\title{Removes rows with any NA values}
\usage{
NoNANA(x)

NoNANA(x)
}
\arguments{
\item{x}{sleuth test table from wald test}
}
\value{



}
\description{
This is just a quicker function to remove any row containing any
NA value. Note this may be overkill in many cases. I use it for
processing proteomic matrices where NA is an indicator of unreliable
measurements.
}
