% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noNANA.R
\name{noNANA}
\alias{noNANA}
\title{Removes rows with any NA values}
\usage{
noNANA(x, margin = "rows")
}
\arguments{
\item{x}{matrix with NA values}

\item{margin}{indicator of what should be done when NA is encountered.
"rows" (default) = delete entire row with NA
"columns" = delete entire column with NA
"cellempty" = replace NA value with empty space
"cellzero" = replace NA value with numeric zero value}
}
\description{
This is just a quicker function to remove any row containing any
NA value. Note this may be overkill in many cases. I use it for
processing proteomic matrices where NA is an indicator of unreliable
measurements.
}
