% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn.smooth.R
\name{syn.smooth}
\alias{syn.smooth}
\title{Title}
\usage{
syn.smooth(in.dir, out.dir = "syn.smooth.results", csv = TRUE, xlsx = TRUE)
}
\arguments{
\item{in.dir}{directory where all input .csv files are located. Omit trailing slash.
i.e. in.dir = "LViN15", or in.dir = "LViN15/results"}

\item{out.dir}{directory where output files should go. Default is
out.dir = "syn.smooth.results"}

\item{csv}{Logical. Indicate whether you want to write .csv files to output directory}

\item{xlsx}{Logical. Indicate whether you want to write .xlsx file with separate sheets
for each output matrix to output directory. NOTE: this uses the xlsx::write.xlsx function
which takes a lot of time and should not be used for very large data sets (larger than
1000 observations)}
}
\description{
Title
}
