% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrimEnrich.R
\name{TrimEnrich}
\alias{TrimEnrich}
\title{Trim mkEnrich() results}
\usage{
TrimEnrich(x, terms = NULL)
}
\arguments{
\item{x}{Output object from mkEnrich() function}

\item{terms}{Terms in pathways to keep. Can use multiple terms separated by pipes.
i.e. "synapse|amyloid|insulin" will subset to include only terms with one of those
names in the description. Not sensitive to case.}
}
\value{
Modified enrichResult only including indicated terms
}
\description{
Many of the enriched pathways from GO enrichment are highly overlapping.
i.e. BC: Ribosome and BP Ribosome Function, etc.
To clean up the downstream Gene concept networks it can be helpful to trim
redundant enriched pathways.
}
